requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pytesseract
pillow
opencv-python
python-dateutil

